Package declaration changes.
