+++ package.json
@@ -12,8 +12,10 @@
   "scripts": {
     "build": "tsc"
   },
-  "main": "index.js",
+  "main": "dist/index.js",
+  "module": "src/index.ts",
   "files": ["dist"],
+  "types": "dist/index.d.ts",
   "dependencies": {
     "ms": "^2.1.3",
     "object-hash": "^2.1.1"
